# DOCUMENTACAO - openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Desafio Técnico - Pulse",
        "version": "1.0",
        "contact": {
            "name": "JOAO HIGO SOUSA NUNES",
            "email": "higo.sousaa@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "local"
        }
    ],
    "paths": {
...
        "/api/cliente/cartao": {
            "get": {
                "summary": "Your GET endpoint",
                "tags": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "description": "",
                                    "minItems": 1,
                                    "uniqueItems": true,
                                    "x-examples": {
                                        "example-1": [
                                            {
                                                "id": 1,
                                                "numero": "23123423423423",
                                                "vencimento": "05/2021",
                                                "cvv": 897
                                            }
                                        ]
                                    },
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "numero": {
                                                "type": "string",
                                                "minLength": 1
                                            },
                                            "vencimento": {
                                                "type": "string",
                                                "minLength": 1
                                            },
                                            "cvv": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "numero": "23123423423423",
                                                "vencimento": "05/2021",
                                                "cvv": 897
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Nao-autorizado"
                    }
                },
                "operationId": "get-api-cliente-cartao",
                "description": "Lista de cartões do cliente",
                "security": [
                    {
                        "Bearer token": []
                    }
                ]
            },
            "post": {
                "summary": "",
                "operationId": "post-api-cliente-cartao",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "",
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "numero": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "vencimento": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "cvv": {
                                            "type": "number"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "numero",
                                        "vencimento",
                                        "cvv"
                                    ],
                                    "x-examples": {
                                        "example-1": {
                                            "id": 2,
                                            "numero": "1515215151255645",
                                            "vencimento": "05/2022",
                                            "cvv": 456
                                        }
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": {
                                            "id": 2,
                                            "numero": "1515215151255645",
                                            "vencimento": "05/2022",
                                            "cvv": 456
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Nao-autorizado"
                    }
                },
                "description": "Inserir Cartão no cadastro do cliente",
                "security": [
                    {
                        "Bearer token": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "",
                                "type": "object",
                                "properties": {
                                    "numero": {
                                        "type": "number"
                                    },
                                    "cvv": {
                                        "type": "number"
                                    },
                                    "vencimento": {
                                        "type": "string",
                                        "minLength": 1
                                    }
                                },
                                "required": [
                                    "numero",
                                    "cvv",
                                    "vencimento"
                                ],
                                "x-examples": {
                                    "example-1": {
                                        "numero": 1515215151255645,
                                        "cvv": 456,
                                        "vencimento": "05/2022"
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "numero": 1515215151255645,
                                        "cvv": 456,
                                        "vencimento": "05/2022"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "",
                "operationId": "put-api-cliente-cartao",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "",
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "numero": {
                                            "type": "number"
                                        },
                                        "cvv": {
                                            "type": "number"
                                        },
                                        "vencimento": {
                                            "type": "string",
                                            "minLength": 1
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "numero",
                                        "cvv",
                                        "vencimento"
                                    ],
                                    "x-examples": {
                                        "example-1": {
                                            "id": 2,
                                            "numero": 1515215151255645,
                                            "cvv": 456,
                                            "vencimento": "05/2025"
                                        }
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": {
                                            "id": 2,
                                            "numero": 1515215151255645,
                                            "cvv": 456,
                                            "vencimento": "05/2025"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Nao-autorizado"
                    },
                    "404": {
                        "$ref": "#/components/responses/Nao-encontrado"
                    }
                },
                "security": [
                    {
                        "Bearer token": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "",
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "numero": {
                                        "type": "number"
                                    },
                                    "cvv": {
                                        "type": "number"
                                    },
                                    "vencimento": {
                                        "type": "string",
                                        "minLength": 1
                                    }
                                },
                                "required": [
                                    "id",
                                    "numero",
                                    "cvv",
                                    "vencimento"
                                ],
                                "x-examples": {
                                    "example-1": {
                                        "id": 2,
                                        "numero": 1515215151255645,
                                        "cvv": 456,
                                        "vencimento": "05/2025"
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "id": 2,
                                        "numero": 1515215151255645,
                                        "cvv": 456,
                                        "vencimento": "05/2025"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer token": {
                "name": "Authorization",
                "type": "apiKey",
                "in": "header",
                "description": ""
            },
            "Perfil": {
                "type": "oauth2"
            }
        },
        "responses": {
            "Nao-autorizado": {
                "description": "Resposta padrão quando um usuário tenta acessar um recurso que ele não tem autorização.",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "",
                            "type": "object",
                            "properties": {
                                "timestamp": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "status": {
                                    "type": "number"
                                },
                                "error": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "message": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "path": {
                                    "type": "string",
                                    "minLength": 1
                                }
                            },
                            "required": [
                                "timestamp",
                                "status",
                                "error",
                                "message",
                                "path"
                            ],
                            "x-examples": {
                                "example-1": {
                                    "timestamp": "2021-05-18T01:24:42.970+00:00",
                                    "status": 403,
                                    "error": "Forbidden",
                                    "message": "Access Denied",
                                    "path": "/api/cliente/cartao"
                                }
                            }
                        },
                        "examples": {
                            "example-1": {
                                "value": {
                                    "timestamp": "2021-05-18T01:24:42.970+00:00",
                                    "status": 403,
                                    "error": "Forbidden",
                                    "message": "Access Denied",
                                    "path": "/api"
                                }
                            }
                        }
                    }
                }
            },
            "Nao-encontrado": {
                "description": "Example response",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "",
                            "type": "object",
                            "properties": {
                                "timeStamp": {
                                    "type": "number"
                                },
                                "status": {
                                    "type": "number"
                                },
                                "error": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "message": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "path": {
                                    "type": "string",
                                    "minLength": 1
                                }
                            },
                            "required": [
                                "timeStamp",
                                "status",
                                "error",
                                "message",
                                "path"
                            ],
                            "x-examples": {
                                "example-1": {
                                    "timeStamp": 1621303164316,
                                    "status": 404,
                                    "error": "Não Encontrado",
                                    "message": "Endereço não encontrado no cadastro do cliente",
                                    "path": "/api/cliente/endereco"
                                }
                            }
                        },
                        "examples": {
                            "example-1": {
                                "value": {
                                    "timeStamp": 1621303164316,
                                    "status": 404,
                                    "error": "Não Encontrado",
                                    "message": "Informação não encontrada no cadastro do cliente",
                                    "path": "/api/"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}